Convert _photon_timings to integer values

+++ fuse/plugins/detector_physics/s2_photon_propagation.py
@@ -392,7 +392,7 @@ class S2PhotonPropagationBase(FuseBaseDownChunkingPlugin):
                     positions,
                     interactions_chunk["sum_s2_photons"],
                     _photon_channels,
-                )
+                ).astype(np.int64)
 
                 # Repeat for n photons per electron # Should this be before adding delays?
                 _photon_timings += np.repeat(electron_group["time"], electron_group["n_s2_photons"])
@@ -465,7 +465,7 @@ class S2PhotonPropagationBase(FuseBaseDownChunkingPlugin):
             positions,
             interactions_chunk["sum_s2_photons"],
             _photon_channels,
-        )
+        ).astype(np.int64)
 
         _photon_timings += np.repeat(electron_group["time"], electron_group["n_s2_photons"])
